requirements.txt
# PaddlePaddle Framework
paddlepaddle-gpu==3.2.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Machine Learning
numpy>=1.19.0
scipy>=1.6.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0
visualdl>=2.2.0

# Progress Bar
tqdm>=4.60.0

# Image Quality Assessment
lpips>=0.1.3
torch>=1.7.0
torchvision>=0.8.0

# Configuration Management
PyYAML>=5.4.0

# Data Augmentation
albumentations>=1.0.0

# Utilities
tensorboard>=2.5.0
wandb>=0.10.0

# Development
pytest>=6.2.0
flake8>=3.9.0
black>=21.0.0

# Optional: for advanced features
# tensorflow>=2.5.0
# keras>=2.5.0